requirements.txt
mysqlclient==2.2.1
djangorestframework 
djangorestframework-simplejwt
python-dotenv
requests